+++ LibSource/FastFourierTransform.cpp
@@ -6,13 +6,13 @@
 #ifdef ARM_CORTEX
 FastFourierTransform::FastFourierTransform(){}
 
-FastFourierTransform::FastFourierTransform(int len){
+FastFourierTransform::FastFourierTransform(size_t len){
   init(len);
 }
 
 FastFourierTransform::~FastFourierTransform(){}
 
-void FastFourierTransform::init(int len){
+void FastFourierTransform::init(size_t len){
   ASSERT(len==32 || len ==64 || len==128 || len==256 || len==512 || len==1024 || len==2048 || len==4096, "Unsupported FFT size");
   void* args[] = {(void*)&instance, (void*)&len};
   getProgramVector()->serviceCall(OWL_SERVICE_ARM_RFFT_FAST_INIT_F32, args, 2);
@@ -32,7 +32,7 @@ void FastFourierTransform::ifft(ComplexFloatArray in, FloatArray out){
   arm_rfft_fast_f32(&instance, (float*)in, (float*)out, 1);
 }
 
-int FastFourierTransform::getSize(){
+size_t FastFourierTransform::getSize(){
   return instance.fftLenRFFT;
 }
 
@@ -40,7 +40,7 @@ int FastFourierTransform::getSize(){
 
 FastFourierTransform::FastFourierTransform(){}
 
-FastFourierTransform::FastFourierTransform(int aSize){
+FastFourierTransform::FastFourierTransform(size_t aSize){
   init(aSize);
 }
 
@@ -48,7 +48,7 @@ FastFourierTransform::~FastFourierTransform(){
   ComplexFloatArray::destroy(temp);
 }
 
-void FastFourierTransform::init(int aSize){
+void FastFourierTransform::init(size_t aSize){
   ASSERT(aSize==32 || aSize ==64 || aSize==128 || aSize==256 || aSize==512 || aSize==1024 || aSize==2048 || aSize==4096, "Unsupported FFT size");
   cfgfft = kiss_fft_alloc(aSize, 0 , 0, 0);
   cfgifft = kiss_fft_alloc(aSize, 1,0, 0);
@@ -58,7 +58,7 @@ void FastFourierTransform::init(int aSize){
 void FastFourierTransform::fft(FloatArray input, ComplexFloatArray output){
   ASSERT(input.getSize() >= getSize(), "Input array too small");
   ASSERT(output.getSize() >= getSize(), "Output array too small");
-  for(int n=0; n<getSize(); n++){
+  for(size_t n=0; n<getSize(); n++){
     temp[n].re=input[n];
     temp[n].im=0;
   }
@@ -70,12 +70,12 @@ void FastFourierTransform::ifft(ComplexFloatArray input, FloatArray output){
   ASSERT(output.getSize() >= getSize(), "Output array too small");
   kiss_fft(cfgifft, (kiss_fft_cpx*)(float*)input, (kiss_fft_cpx*)(float*)temp.getData());
   float scale=1.0f/getSize();
-  for(int n=0; n<getSize(); n++){
+  for(size_t n=0; n<getSize(); n++){
     output[n]=temp[n].re*scale;
   }
 }
     
-int FastFourierTransform::getSize(){
+size_t FastFourierTransform::getSize(){
   return temp.getSize();
 }
